Move expand button to expanders dir
